requirements.txt
streamlit
Pillow
rembg
onnxruntime